O: rename `utilities.mjs` to `types.mjs` O: use `types.isString()` for checking string or not O: do not check `instanceof String`
